moved model build to builder.py
